지울 파일명 전송

+++ src/main/java/org/zerock/service/board/BoardSerivce.java
@@ -91,26 +91,28 @@ public class BoardSerivce {
 	public int update(BoardDto board, MultipartFile[] files) {
 		
 		for (MultipartFile file : files) {
-			int boardId = board.getId();
-			String name = file.getOriginalFilename();
-			// File table에 해당파일명 지우기
-			boardMapper.deleteFileByBoardIdAndFileName(boardId, name);
-			
-			// File table에 파일명 추가
-			boardMapper.insertFile(boardId, name);
-			
-			// 저장소에 실제 파일 추가
-			File folder = new File("C:\\Users\\user\\Desktop\\study\\upload\\prj1\\board\\" + board.getId());
-			folder.mkdirs();
-			
-			File dest = new File(folder, name);
-			
-			try {
-				file.transferTo(dest);
-			} catch (Exception e) {
-				// @Transactional은 RuntimeException에서만 rollback 됨
-				e.printStackTrace();
-				throw new RuntimeException(e);
+			if (file != null && file.getSize() > 0) {
+				int boardId = board.getId();
+				String name = file.getOriginalFilename();
+				// File table에 해당파일명 지우기
+				boardMapper.deleteFileByBoardIdAndFileName(boardId, name);
+				
+				// File table에 파일명 추가
+				boardMapper.insertFile(boardId, name);
+				
+				// 저장소에 실제 파일 추가
+				File folder = new File("C:\\Users\\user\\Desktop\\study\\upload\\prj1\\board\\" + board.getId());
+				folder.mkdirs();
+				
+				File dest = new File(folder, name);
+				
+				try {
+					file.transferTo(dest);
+				} catch (Exception e) {
+					// @Transactional은 RuntimeException에서만 rollback 됨
+					e.printStackTrace();
+					throw new RuntimeException(e);
+				}
 			}
 			
 		}
